+++ src/components/ExhibitorList/index.js
@@ -14,7 +14,6 @@ import { GrCheckbox, GrCheckboxSelected } from 'react-icons/gr'
 import { BsSliders } from 'react-icons/bs'
 import { PlaceGoldFirst } from '@/templates/placeGoldFirst'
 import { FaExternalLinkAlt } from 'react-icons/fa'
-import Fuse from 'fuse.js'
 
 /* armada.nu/exhibitors is no longer being used. To do is to patch all this and make it work with the API again //Nima
 
@@ -28,172 +27,6 @@ import Fuse from 'fuse.js'
 }*/
 
 //base of server adress
-
-const options = {
-    keys: ['program'], // Search based on the program name
-    threshold: 0.1, // Adjust the threshold for fuzzy matching
-}
-
-const programsAndIndustries = [
-    {
-        program: 'Biomedical Engineering',
-        industries: ['Pharmaceutical', 'Biotechnology'],
-    },
-    {
-        program: 'Chemical Engineering',
-        industries: [
-            'Solid Mechanics',
-            'Pharmaceutical',
-            'Biotechnology',
-            'Nuclear Power',
-            'Energy Technology',
-            'Nanotechnology',
-        ],
-    },
-    {
-        program: 'Civil Engineering',
-        industries: [
-            'Architecture',
-            'Construction',
-            'Property & Infrastructure',
-            'Railway',
-        ],
-    },
-    {
-        program: 'Computer Science & Engineering',
-        industries: [
-            'Simulation Technology',
-            'Software Development',
-            'Web Development',
-            'Telecommunication',
-            'IT Infrastructure',
-            'Interaction Design',
-        ],
-    },
-    {
-        program: 'Electrical Engineering',
-        industries: [
-            'Acoustics',
-            'Aerospace',
-            'Telecommunication',
-            'Mechatronics',
-            'Electronics',
-            'Nanotechnology',
-        ],
-    },
-    {
-        program: 'Engineering Mathematics & Physics',
-        industries: [
-            'Solid Mechanics',
-            'Acoustics',
-            'Nuclear Power',
-            'Fluid Mechanics',
-            'Industry Design',
-        ],
-    },
-    {
-        program: 'Environmental Engineering',
-        industries: ['Environmental Sector', 'Energy Technology'],
-    },
-    {
-        program: 'Industrial Engineering',
-        industries: [
-            'Manufacturing Industry',
-            'Management Consulting',
-            'Insurance',
-            'Finance',
-            'Logistics & Supply Chain',
-            'Mechatronics',
-            'Property & Infrastructure',
-            'Industry Design',
-            'Recruitment',
-        ],
-    },
-    {
-        program: 'Information Technology',
-        industries: [
-            'Web Development',
-            'Simulation Technology',
-            'Telecommunication',
-            'IT Infrastructure',
-            'Software Development',
-            'Interaction Design',
-        ],
-    },
-    {
-        program: 'Mechanical Engineering',
-        industries: [
-            'Automotive',
-            'Fluid Mechanics',
-            'Solid Mechanics',
-            'Aerospace',
-            'Acoustics',
-            'Marine System',
-            'Mechatronics',
-            'Nanotechnology',
-        ],
-    },
-    {
-        program: 'Media Technology',
-        industries: [
-            'Web Development',
-            'Simulation Technology',
-            'Media Technology',
-            'Telecommunication',
-            'IT Infrastructure',
-            'Software Development',
-            'Interaction Design',
-        ],
-    },
-    {
-        program: 'Medical Engineering',
-        industries: ['Medical Technology'],
-    },
-    {
-        program: 'Material & Product Design',
-        industries: [
-            'Pharmaceutical',
-            'Wood-Processing Industry',
-            'Steel Industry',
-            'Manufacturing Industry',
-            'Logistics & Supply Chain',
-            'Material Development',
-            'Nanotechnology',
-            'Product Development',
-        ],
-    },
-    {
-        program: 'Other',
-        industries: ['Research', 'Pedagogy', 'Retail'],
-    },
-    // Add more program-industry associations
-]
-
-const fuse = new Fuse(programsAndIndustries, options) //instance for fuzzy searching
-
-function matchProgramToIndustries(program) {
-    const results = fuse.search(program)
-    const industries = results.map(result => result.item.industries).flat()
-    return industries
-}
-
-function matchIndustriesToExhibitors(industries, allExhibitors) {
-    return allExhibitors.reduce((matchedExhibitors, exhibitor) => {
-        // Check if the exhibitor has all the specified industries
-        const hasAllIndustries = industries.some(industry =>
-            exhibitor.industries.some(
-                exhibitorIndustry => exhibitorIndustry.name === industry
-            )
-        )
-
-        // If the exhibitor has all the specified industries, add it to the result array
-        if (hasAllIndustries) {
-            matchedExhibitors.push(exhibitor)
-        }
-
-        return matchedExhibitors
-    }, [])
-}
 
 const dropDownAttributes = {
     width: '100%',
@@ -261,6 +94,7 @@ export class ExhibitorList extends React.Component {
                 .toString(), //get the previous year
             exhibitors: [], // json object
             exhibitorList: [], //displayed exhibitors
+            recommendedExhibitors: [],
             showModal: false, //show individual company card
             exhibitorName: undefined,
             isLoading: true,
@@ -661,9 +495,26 @@ export class ExhibitorList extends React.Component {
         })
     }
 
+    updateRecommendedExhibitors(recommendedExhibitorsInput) {
+        let recommendedExhibitorsList = recommendedExhibitorsInput.map(
+            exhibitor => (
+                <ExhibitorItem
+                    key={exhibitor.id}
+                    name={exhibitor.name}
+                    exhibitor={exhibitor}
+                    showModal={this.showModal}
+                />
+            )
+        )
+        this.setState({
+            recommendedExhibitors: recommendedExhibitorsList,
+        })
+    }
+
     componentDidUpdate(prevProps) {
         //console.log(this.props.fairInputLocation, prevProps.fairInputLocation) //remove this later
         //console.log(this.props.fairInputExhibitors) //remove this also later.
+
         if (this.props.fairInputLocation !== prevProps.fairInputLocation) {
             this.updateLocationShowed(this.props.fairInputLocation)
         }
@@ -678,6 +529,12 @@ export class ExhibitorList extends React.Component {
                 exhibitorName: this.props.exhibitorName,
                 showModal: true,
             })
+        }
+        if (
+            JSON.stringify(this.props.recommendedExhibitors) !==
+            JSON.stringify(prevProps.recommendedExhibitors)
+        ) {
+            this.updateRecommendedExhibitors(this.props.recommendedExhibitors)
         }
 
         // Do not update unless the floor changed
@@ -1089,7 +946,6 @@ export class ExhibitorList extends React.Component {
         // Here you decide if list of exhibitors should be displayed or not
         let showExhibitors = true
         let thisYear = new Date().getFullYear().toString()
-        //console.log('poop' + this.props.fairLocation)
         if (this.year === thisYear) {
             showExhibitors = false
         }
@@ -1257,34 +1113,6 @@ export class ExhibitorList extends React.Component {
         if (showExhibitors) {
             return (
                 <div className='exhibitors'>
-                    {/*<Questionnaire />*/}
-                    {/* <h1>
-                        {this.props.lastYear ? "Last Year's " : ''}Exhibitors
-                    </h1>
-                    <br />
-                    <p
-                        style={{
-                            width: '90%',
-                            margin:'0em 2em 0em 2em',
-                            paddingBottom: this.props.lastYear ? '1em' : {},
-                        }}
-                    >
-                        {this.props.lastYear ? (
-                            `These are the exhibitors from the ${this.state.previousYear} fair.`
-                        ) : (
-                            <span>
-                                <span className='bold'>
-                                    Sustainability & Diversity
-                                </span>{' '}
-                                form the core values at the heart of our
-                                organization. To highlight our core values, we
-                                have chosen to dedicate focus areas of the fair
-                                called Green Room and Diversity Room. If an
-                                exhibitor is tagged with one of the images
-                                below, they are in one of these rooms!
-                            </span>
-                        )}
-                    </p> */}
                     {this.state.showModal
                         ? this.displayExhibitor(exhibitorToDisplay)
                         : null}
@@ -1487,14 +1315,16 @@ export class ExhibitorList extends React.Component {
                             (this.state.showModal ? 'notDisplay' : '')
                         }
                     >
-                        <div className='recommended-exhibitors'>
-                            <div className='recommended-exhibitors-text'>
-                                <b>Recommended exhibitors for you</b>
+                        {this.state.recommendedExhibitors > 0 && (
+                            <div className='recommended-exhibitors'>
+                                <div className='recommended-exhibitors-text'>
+                                    <b>Recommended exhibitors for you</b>
+                                </div>
+                                <div className='recommended-exhibitors-entries'>
+                                    {this.state.recommendedExhibitors}
+                                </div>
                             </div>
-                            <div className='recommended-exhibitors-entries'>
-                                {}
-                            </div>
-                        </div>
+                        )}
                         <div className='all-exhibitors'>
                             <div className='all-exhibitors-text'>
                                 <b>All Exhibitors</b>
@@ -1552,6 +1382,7 @@ ExhibitorList.propTypes = {
     lastYear: PropTypes.bool,
     year: PropTypes.string,
     showCV: PropTypes.bool,
+    recommendedExhibitors: PropTypes.array,
 }
 
 export default ExhibitorList
