+++ src/components/Questionnaire/index.js
@@ -6,8 +6,9 @@ import 'survey-core/defaultV2.min.css'
 import './index.scss'
 import { surveyLocalization } from 'survey-core'
 import './pop.scss'
+import Fuse from 'fuse.js'
 
-const Questionnaire = ({ setShowButtons }) => {
+const Questionnaire = ({ setShowButtons, setRecommendedExhibitors }) => {
     const majorList = [
         'Biomedical Engineering',
         'Chemical Engineering',
@@ -34,6 +35,172 @@ const Questionnaire = ({ setShowButtons }) => {
         'Bachelor thesis',
         'Master thesis',
     ]
+
+    const programsAndIndustries = [
+        {
+            program: 'Biomedical Engineering',
+            industries: ['Pharmaceutical', 'Biotechnology'],
+        },
+        {
+            program: 'Chemical Engineering',
+            industries: [
+                'Solid Mechanics',
+                'Pharmaceutical',
+                'Biotechnology',
+                'Nuclear Power',
+                'Energy Technology',
+                'Nanotechnology',
+            ],
+        },
+        {
+            program: 'Civil Engineering',
+            industries: [
+                'Architecture',
+                'Construction',
+                'Property & Infrastructure',
+                'Railway',
+            ],
+        },
+        {
+            program: 'Computer Science & Engineering',
+            industries: [
+                'Simulation Technology',
+                'Software Development',
+                'Web Development',
+                'Telecommunication',
+                'IT Infrastructure',
+                'Interaction Design',
+            ],
+        },
+        {
+            program: 'Electrical Engineering',
+            industries: [
+                'Acoustics',
+                'Aerospace',
+                'Telecommunication',
+                'Mechatronics',
+                'Electronics',
+                'Nanotechnology',
+            ],
+        },
+        {
+            program: 'Engineering Mathematics & Physics',
+            industries: [
+                'Solid Mechanics',
+                'Acoustics',
+                'Nuclear Power',
+                'Fluid Mechanics',
+                'Industry Design',
+            ],
+        },
+        {
+            program: 'Environmental Engineering',
+            industries: ['Environmental Sector', 'Energy Technology'],
+        },
+        {
+            program: 'Industrial Engineering',
+            industries: [
+                'Manufacturing Industry',
+                'Management Consulting',
+                'Insurance',
+                'Finance',
+                'Logistics & Supply Chain',
+                'Mechatronics',
+                'Property & Infrastructure',
+                'Industry Design',
+                'Recruitment',
+            ],
+        },
+        {
+            program: 'Information Technology',
+            industries: [
+                'Web Development',
+                'Simulation Technology',
+                'Telecommunication',
+                'IT Infrastructure',
+                'Software Development',
+                'Interaction Design',
+            ],
+        },
+        {
+            program: 'Mechanical Engineering',
+            industries: [
+                'Automotive',
+                'Fluid Mechanics',
+                'Solid Mechanics',
+                'Aerospace',
+                'Acoustics',
+                'Marine System',
+                'Mechatronics',
+                'Nanotechnology',
+            ],
+        },
+        {
+            program: 'Media Technology',
+            industries: [
+                'Web Development',
+                'Simulation Technology',
+                'Media Technology',
+                'Telecommunication',
+                'IT Infrastructure',
+                'Software Development',
+                'Interaction Design',
+            ],
+        },
+        {
+            program: 'Medical Engineering',
+            industries: ['Medical Technology'],
+        },
+        {
+            program: 'Material & Product Design',
+            industries: [
+                'Pharmaceutical',
+                'Wood-Processing Industry',
+                'Steel Industry',
+                'Manufacturing Industry',
+                'Logistics & Supply Chain',
+                'Material Development',
+                'Nanotechnology',
+                'Product Development',
+            ],
+        },
+        {
+            program: 'Other',
+            industries: ['Research', 'Pedagogy', 'Retail'],
+        },
+        // Add more program-industry associations
+    ]
+
+    const options = {
+        keys: ['program'], // Search based on the program name
+        threshold: 0.1, // Adjust the threshold for fuzzy matching
+    }
+
+    const fuse = new Fuse(programsAndIndustries, options) //instance for fuzzy searching
+
+    function matchProgramToIndustries(program) {
+        const results = fuse.search(program)
+        const industries = results.map(result => result.item.industries).flat()
+        return industries
+    }
+
+    function matchIndustriesToExhibitors(industries, allExhibitors) {
+        return allExhibitors.reduce((matchedExhibitors, exhibitor) => {
+            // Check if the exhibitor has all the specified industries
+            const hasAllIndustries = industries.some(industry =>
+                exhibitor.industries.some(
+                    exhibitorIndustry => exhibitorIndustry.name === industry
+                )
+            )
+
+            // If the exhibitor has all the specified industries, add it to the result array
+            if (hasAllIndustries) {
+                matchedExhibitors.push(exhibitor)
+            }
+
+            return matchedExhibitors
+        }, [])
+    }
 
     // const options = {
     //     keys: ['program'], // Search based on the program name
